Add database for syntax sugar
