+++ crates/analyzer/src/scope.rs
@@ -1,7 +1,6 @@
 //! Database for scope information.
 mod collect;
 mod data;
-mod recursive;
 mod resolve;
 
 use crate::{id::InFile, resolver::ValueGroupId, SurfaceDatabase};
@@ -9,8 +8,6 @@ use std::sync::Arc;
 
 use collect::CollectContext;
 pub use data::*;
-use files::FileId;
-use recursive::{BindingGroupsContext, LetBindingGroupsContext};
 use resolve::ResolveContext;
 
 #[salsa::query_group(ScopeStorage)]
@@ -20,10 +17,4 @@ pub trait ScopeDatabase: SurfaceDatabase {
 
     #[salsa::invoke(ResolveContext::value_resolve_query)]
     fn value_resolved(&self, id: InFile<ValueGroupId>) -> Arc<ValueGroupResolutions>;
-
-    #[salsa::invoke(BindingGroupsContext::binding_groups_query)]
-    fn binding_groups(&self, file_id: FileId) -> Arc<BindingGroups>;
-
-    #[salsa::invoke(LetBindingGroupsContext::let_binding_groups_qiery)]
-    fn let_binding_groups(&self, id: InFile<ValueGroupId>) -> Arc<LetBindingGroups>;
 }
